test(onboard): cover gateway cleanup trust boundaries
